+++ lib/djv/Models/SettingsModel.cpp
@@ -109,7 +109,6 @@ namespace djv
             std::shared_ptr<ftk::Settings> settings;
             std::map<std::string, Shortcut> savedShortcuts;
             ShortcutsSettings shortcutsDefault;
-            float displayScaleDefault = 1.F;
 
             std::shared_ptr<ftk::Observable<AudioSettings> > audio;
             std::shared_ptr<ftk::Observable<tl::PlayerCacheOptions> > cache;
@@ -162,14 +161,12 @@ namespace djv
 
         void SettingsModel::_init(
             const std::shared_ptr<ftk::Context>& context,
-            const std::shared_ptr<ftk::Settings>& settings,
-            float displayScaleDefault)
+            const std::shared_ptr<ftk::Settings>& settings)
         {
             FTK_P();
 
             p.context = context;
             p.settings = settings;
-            p.displayScaleDefault = displayScaleDefault;
 
             AudioSettings audio;
             settings->getT(keys["Audio"], audio);
@@ -274,7 +271,6 @@ namespace djv
             p.playback = ftk::Observable<PlaybackSettings>::create(playback);
 
             StyleSettings style;
-            style.displayScale = displayScaleDefault;
             settings->getT(keys["Style"], style);
             p.style = ftk::Observable<StyleSettings>::create(style);
 
@@ -316,11 +312,10 @@ namespace djv
 
         std::shared_ptr<SettingsModel> SettingsModel::create(
             const std::shared_ptr<ftk::Context>& context,
-            const std::shared_ptr<ftk::Settings>& settings,
-            float displayScaleDefault)
+            const std::shared_ptr<ftk::Settings>& settings)
         {
             auto out = std::shared_ptr<SettingsModel>(new SettingsModel);
-            out->_init(context, settings, displayScaleDefault);
+            out->_init(context, settings);
             return out;
         }
 
@@ -407,9 +402,7 @@ namespace djv
             setMisc(miscSettings);
             setMouse(MouseSettings());
             setPlayback(PlaybackSettings());
-            StyleSettings style;
-            style.displayScale = p.displayScaleDefault;
-            setStyle(style);
+            setStyle(StyleSettings());
             setTimeline(TimelineSettings());
             setWindow(WindowSettings());
 #if defined(TLRENDER_FFMPEG_PLUGIN)